backend-suzan/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-pinecone>=0.1.1
pinecone-client==3.2.2
pypdf==4.2.0
requests==2.32.3
sqlmodel==0.0.16
sqlalchemy==2.0.29
aiosqlite==0.19.0
pydantic==2.7.4
pydantic-settings
better-profanity==0.7.0
sentry-sdk[fastapi]
langsmith
psycopg2-binary==2.9.9
